Core functionality for experiment one added
Issues with experiment one - can't seem to get the same results as the paper. Failure functionality appears completely random. Attack has some sort of consistency. Issue is with the size of the graph and the performance hardware can give. It is stated clearly within the paper that results should be seen across all graphs. Need to look at <<k>> and other ways of finding the diameter/randomly removing nodes in a graph.